Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

this, combined with the change to proprietary_vendor_samsung/aries-common resulted in a null interface being passed to the camera application. The camera recognizes that it has certain features and resolutions, but can never connect to the hardware it seems (or even preview the image). Clearing Camera's application data did not fix this.

http://pastebin.com/ziL8giSP

Patch-set: 1