Update patch set 1
Patch Set 1:

I'm sorry, I'm having a hard time understanding what it is you wrote. The problem isn't that cache partition is actually too small, it's that the dalvik-cache files take up a huge amount of space on this partition. What's the point in moving the location of the downloaded file instead of the dalvik cache for system/app when we already have this in place?

Going back to a conversation that was had earlier in our IRC channel, why is it that we don't released odex'ed "user" builds instead of engineering/deodexed builds? it would cut down on first boot time as well as not waste any space in /cache for dexopt'ing.

Patch-set: 1