Update patch set 1
Patch Set 1: I would prefer that you didn't submit this

You're removing the kernel and modules without leaving any sort of fallback in case the kernel source can't be found (which is the whole point of the redundant checking of source). Right now, if source is for some reason unavailable, you cannot build with this change.

TARGET_PREBUILT_KERNEL needs to be defined on a per-device basis.

Patch-set: 1